Improve QoL for OSS users of FBShipIt (#100)
Summary:
Pull Request resolved: https://github.com/facebook/fbshipit/pull/100

This is a minor refactor that makes it a little easier to work with the public version of shipit.

I found when writing the demo that it was odd that there was no built-in filter to do this, but I realize now that's because in the OSS version it's tucked away in fb-examples and I missed it.

Reviewed By: yns88

Differential Revision: D21642531

fbshipit-source-id: b25286d6b06626b12a3600594e4b0d557b185565