s/HH_IGNORE_ERROR/HH_FIXME
Summary:
We're going to ban HH_IGNORE_ERROR in order to unify the methods of suppressing errors. This is a codemod that straight converts one to the other and then deals with a metric ton of codegen.

All tests are unrelated / already failing in trunk. The other failures are caused by issues rebasing. I'll drop conflicts and stuff.

Reviewed By: xulingc

Differential Revision: D21672246

fbshipit-source-id: bda6b719f47aa3616efda3d62f18a978f5f0b1b4