fix 'Finished phase' logging for phases that throw
Summary: some phases like verify and createrepo throw even on success, to prevent the rest of the sync. but this prevents it from logging the "Finished phase" message. in `--verbose` mode, that message includes a timestamp and it'd be helpful to know how long it took for these phases to run.

Reviewed By: bigfootjon

Differential Revision: D21905459

fbshipit-source-id: b84d2425ad20512479d9ff880eff90101515f052