Link diff IDs to OSS commits in changesets
Summary:
In projects that use ShipIt, it can be hard to read the commit log as a non-FB employee. Often, employees will cite diffs inline in commit messages (i.e. "this solves problem X created by D123"). When exported, it is *possible* to query the repo history to find the "Differential Revision: D123" line but it would be really nice to just link to the commit directly.

This only handles intra-repo commits. Handling inter-repo commit links is not worth the extra complexity.

Reviewed By: jurajh-fb

Differential Revision: D22096330

fbshipit-source-id: 5cc753afb07b3f1316f9402fbcd3ac7fe17fdc52