Optimize project list by filtering on source roots
Summary:
To render a shipitpreview (either inline in the diffs tool or in the bespoke endpoint) each config must be considered.

The way we currently do that is by iterating over each project and applying its `filterChangeset` function. If the changeset returned by this function is empty, then the preview does not render a patch.

`filterChangeset` is an expensive operation, and we can make it much more efficient by first checking the project's sourceroots and filtering on those.

Reviewed By: yns88

Differential Revision: D22140675

fbshipit-source-id: a75827799766b9e9ed318279e0634dc36374eea6