Encode MIME sequences when committing to git
Summary:
When ImportIt consumed [this PyTorch diff](https://github.com/pytorch/pytorch/pull/38440) it persisted the `=?` sequence into the diff title in phabricator. This was able to be committed to mercurial just fine but it broke when ShipIt attempted to commit the patch into the PyTorch repo.

We can fix this with a PHP builtin: the `mb_encode_mimeheader` function.

Reviewed By: yns88

Differential Revision: D22412496

fbshipit-source-id: 1b9a916fb592d2e059de65b7b93cf51ead2a17d6