Fix lock releasing when never acquired
Summary:
There are certain cases where a lock is never actually acquired, for example the destination lock is never acquired when in `--create-new-repo` mode.

Problem is, the way the lock releasing code works is that if there isn't a lock ShipIt will *acquire* the lock -- only to immediately release it. Instead, we should just never acquire the lock.

I knew this was theoretically possible when I implemented it this way but my testing didn't raise any issues.

This is [causing problems for real use-cases](https://fb.workplace.com/groups/osssupport/permalink/4159409230774327/) so let's fix it.

Reviewed By: yns88

Differential Revision: D23109040

fbshipit-source-id: 3121a3b928982843e770fe6ad9c4b56dcedc1fe0