Print out the patch that failed to sync
Summary:
This originally came from a request from Wez, but I hit this when debugging a react-native issue.

When run from the command-line, the stack trace includes the patch, but when run in other contexts (like within a web request or CI job) the stack trace is not printed so it's unclear what went wrong. This prints out the patch no matter the context.

Reviewed By: jurajh-fb

Differential Revision: D23350421

fbshipit-source-id: 611df04496ff1523ce695ea63c4cba2c8fe474eb