Support creating/verifying repos that are >2GB
Summary:
We've been hitting the upper bound on what ShipIt can handle in a few repos. Generally speaking, things are fine because incremental sync works. The problem comes when a project gets out of sync; in order to re-sync we need to get a clean copy of the repo and compare them. We use the --create-new-repo mechanism to do this, and that is where the real problem lies.

This diff offers a solution: split the initial commit into a series of commits by chunking the files into groups. After chunking, filtering, and creating the new repo then we squash the commit log into a single commit with ShipIt tracking information.

Bam, 2GB limit averted.

Reviewed By: yns88

Differential Revision: D24352978

fbshipit-source-id: 20cefe3f3a409acc7538db155ce512208f6093ac