Syncing Thrift changes from D27088410
Summary:
This is an auto-generated diff that syncs changes in .thrift files from D27088410.

D27088410 modified thrift files that are being synchronized to WWW repository, which triggered creation of this diff.

Please, make sure that all tests pass and accept this diff. It will be updated and landed once the primary diff lands.
If the change requires manual changes, then you should commandeer the diff and fix it (you will be required to land it manually).

You can find a command for regenerating the changes locally and the errors (if any) in the signals section.

You can read more about auto-sync on [the wiki](https://fburl.com/thrift-www-auto-sync-wiki).

**Read Signals section below which contains details about specific problems with the diff**
**Please check out the [common issues](https://fburl.com/thrift-auto-sync-troubleshoot) listed here before posting in the users group.**

Depends on D27088410

<!-----DO NOT EDIT CONTENT BELOW THIS LINE-----!>
<<PARTIALLY AUTOGENERATED FROM D27088410>>

Reviewed By: yns88

Differential Revision: D27090075

fbshipit-source-id: 054b9126cc9edf9ea2991c4c8b368ee5ae323e94