Migrate simple cases of shell commands
Summary:
Ok so this is going to be a big stack. This gist here is that I'm going ahead and async-ifying the entirety of ShipIt. Eventually we will migrate to the HSL (once it's available for shell execs) which will mean 2 changes:
1. Mandatory async
2. We can remove all ShipIt commands from the phps server-mode blacklist

(2) is a really nice performance boost, and also allows us to FINALLY use a debugger with ShipIt.

Anyway, this is a tall stack that basically does nothing besides adding `async`, `await` and `gen` prefixes to functions.

NOTE: It is *NOT* a performance hit to `await` the new `genRun` because `await`s are eagerly evaluated until they reach a point that is ACTUALLY an async call. Since `genRun` just calls `runSynchrounously` (a sync function) the `await`s are immediately evaluated (the exact same as before this diff)

This stack will have MANY diffs in it. I'm sorry.

Reviewed By: 2d2d2d2d2d

Differential Revision: D30253737

fbshipit-source-id: e86d7f27141c98e44ce2eb98310f035603d02aed