Use full git SHAs (#136)
Summary:
Pull Request resolved: https://github.com/facebook/fbshipit/pull/136

Right now mercurial SHAs are always full-length, but git shas are repo-dependent when used as a ShipIt source. `git log --oneline` picks the length of the SHA it renders by the number of commits. This is nice for readability/users but a headache for automation that attempts to look for the tracking SHAs that ShipIt produces

Reviewed By: yns88

Differential Revision: D34693603

fbshipit-source-id: 34033a184602a4853b4970a2c134d18bf5437711