Enable verification of fbsource destination repo
Summary:
This was a pickle, we need to add verification support to a phase that expects the destination will only ever be a git repo.

I... just exported the HG repo into a Git repo using ShipIt, and then used that "fake" destination repo as the comparison repo. It works? I kinda hate that it works?

Reviewed By: yns88

Differential Revision: D34978066

fbshipit-source-id: 63720b65a00a8bae477ddc775dcae3bcae3d8a52