Fix sync running out-of-order
Summary:
In an internal plugin, we have logic that depends on each changeset being filtered in the order it landed.

This change breaks that plugin but ONLY when multiple commits are landed simultaneously. This happens only once per day for that repo so it took me a bit to find the problem.

This is effectively a manually-applied partial revert of D29615024 (https://github.com/facebook/fbshipit/commit/f44259b53c23e0e06a6963c1abfcf60f2a045261) only for this section.

Reviewed By: albanD

Differential Revision: D36005842

fbshipit-source-id: d5d6c701afaed75ffd4241c06cff9948bb953fa9