Checkout HEAD before generating diff in Verify phase
Summary:
# Summary

`shipit-verify-diff-train-pytorch-pytorch` jobs have been very flaky. The root cause of this flakiness is that during `
Verify synchronicity with ShipIt` step, the source git repo is somehow not in the correct state (empty). This results in a diff-patch where the entire repository is being deleted. Subsequently, we fail to apply this huge patch in `Apply Patch to fbsource` step.

Sample jobs with this failure. Check logs of `Verify synchronicity with ShipIt` to see the huge patch:
- https://www.internalfb.com/intern/sandcastle/job/9007199957860250/insights
- https://www.internalfb.com/intern/sandcastle/job/9007199957872394/insights
- https://www.internalfb.com/intern/sandcastle/job/708391496/insights

As part of this change, we simply add a `git checkout HEAD` before we run the diff-patch. This command appears to be correctly restoring the source repo state, and remove the flakiness.

Reviewed By: bigfootjon

Differential Revision: D37390423

fbshipit-source-id: 9a69f82d7fe55c59ee67c7c6ee9a5cd9fb87f18d