Convert HH_IGNORE_ERROR to HH_FIXME[xxxx]
Summary:
We are deprecating the IgnoreError-style comments, so replace them with vanilla HH_FIXME suppressions which they are equivalent to.

This does not determine if a comment is no longer suppressing an active error (i.e. dead fixmes).

I am not regenerating typed loggers since static config generation is broken.

Reviewed By: rodmk

Differential Revision: D38409393

fbshipit-source-id: ceaf92aa20fd91e05d53e2d5a2d22200e0f9a814